requirements.txt
typer==0.12.3
PyYAML==6.0.2
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
tzdata==2025.1
